ChunkedArray class

A ChunkedArray is a data structure managing a list of primitive Arrow Arrays logically as one large array. Chunked arrays may be grouped together in a Table.

chunked_array(..., type = NULL)

Arguments

...
Vectors to coerce

type
currently ignored

Factory

The ChunkedArray$create() factory method instantiates the object from various Arrays or R vectors. chunked_array() is an alias for it.

Methods

$length(): Size in the number of elements this array contains

$chunk(i): Extract an Array chunk by integer position

$as_vector(): convert to an R vector

$Slice(offset, length = NULL): Construct a zero-copy slice of the array with the indicated offset and length. If length is NULL, the slice goes until the end of the array.

$Take(i): return a ChunkedArray with values at positions given by integers i. If i is an Arrow Array or ChunkedArray, it will be coerced to an R vector before taking.

$Filter(i): return a ChunkedArray with values at positions where logical vector or Arrow boolean-type (Chunked)Array i is TRUE.

$cast(target_type, safe = TRUE, options = cast_options(safe)): Alter the data in the array to change its type.

$null_count(): The number of null entries in the array

$chunks(): return a list of Arrays

$num_chunks(): integer number of chunks in the ChunkedArray

$type(): logical type of data

$View(type): Construct a zero-copy view of this ChunkedArray with the given type.

$Validate(): Perform any validation checks to determine obvious inconsistencies within the array's internal data. This can be an expensive check, potentially O(length)

See also

Super class

arrow::Object -> ChunkedArray

Methods

Public methods

ChunkedArray$length()

ChunkedArray$chunk()

ChunkedArray$as_vector()

ChunkedArray$Slice()

ChunkedArray$Take()

ChunkedArray$Filter()

ChunkedArray$cast()

ChunkedArray$View()

ChunkedArray$Validate()

ChunkedArray$ToString()

ChunkedArray$Equals()

ChunkedArray$clone()

Inherited methods

arrow::Object$initialize()

arrow::Object$pointer()

arrow::Object$print()

arrow::Object$set_pointer()

Method length()

Usage

ChunkedArray$length()

Method chunk()

Usage

ChunkedArray$chunk(i)

Method as_vector()

Usage

ChunkedArray$as_vector()

Method Slice()

Usage

ChunkedArray$Slice(offset, length = NULL)

Method Take()

Usage

ChunkedArray$Take(i)

Method Filter()

Usage

ChunkedArray$Filter(i)

Method cast()

Usage

ChunkedArray$cast(target_type, safe = TRUE, options = cast_options(safe))

Method View()

Usage

ChunkedArray$View(type)

Method Validate()

Usage

ChunkedArray$Validate()

Method ToString()

Usage

ChunkedArray$ToString()

Method Equals()

Usage

ChunkedArray$Equals(other)

Method clone()

The objects of this class are cloneable with this method.

Usage

ChunkedArray$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.