FileSystem entry stats

FileSystem entry stats

FileSystem entry stats

Methods

base_name() : The file base name (component after the last directory separator).

extension() : The file extension

Active bindings

$type: The file type

$path: The full file path in the filesystem

$size: The size in bytes, if available. Only regular files are guaranteed to have a size.

$mtime: The time of last modification, if available.

Super class

arrow::Object -> FileStats

Methods

Public methods

FileStats$base_name()

FileStats$extension()

FileStats$clone()

Inherited methods

arrow::Object$initialize()

arrow::Object$pointer()

arrow::Object$print()

arrow::Object$set_pointer()

Method base_name()

Usage

FileStats$base_name()

Method extension()

Usage

FileStats$extension()

Method clone()

The objects of this class are cloneable with this method.

Usage

FileStats$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.