InputStream classes

RandomAccessFile inherits from InputStream and is a base class for: ReadableFile for reading from a file; MemoryMappedFile for the same but with memory mapping; and BufferReader for reading from a buffer. Use these with the various table readers.

Factory

The $create() factory methods instantiate the InputStream object and take the following arguments, depending on the subclass:

path For ReadableFile, a character file name

x For BufferReader, a Buffer or an object that can be made into a buffer via buffer().

To instantiate a MemoryMappedFile, call mmap_open().

Super classes

arrow::Object -> arrow::Readable -> InputStream

arrow::Object -> arrow::Readable -> arrow::InputStream -> RandomAccessFile

arrow::Object -> arrow::Readable -> arrow::InputStream -> arrow::RandomAccessFile -> MemoryMappedFile

arrow::Object -> arrow::Readable -> arrow::InputStream -> arrow::RandomAccessFile -> BufferReader

arrow::Object -> arrow::Readable -> arrow::InputStream -> arrow::RandomAccessFile -> ReadableFile

Methods

Public methods

$close()

$clone()

$Read()

$ReadAt()

$GetSize()

$supports_zero_copy()

$seek()

$tell()

$Resize()

Method close()

Usage

InputStream$close()

RandomAccessFile$close()

MemoryMappedFile$close()

BufferReader$close()

ReadableFile$close()

  

Method clone()

The objects of these classes are cloneable with this method.

Usage

InputStream$clone(deep = FALSE)

RandomAccessFile$clone(deep = FALSE)

MemoryMappedFile$clone(deep = FALSE)

BufferReader$clone(deep = FALSE)

ReadableFile$clone(deep = FALSE)

  

Arguments

deep

Whether to make a deep clone.

Method Read()

Read data from the stream

Usage

InputStream$Read(nbytes = NULL)

RandomAccessFile$Read(nbytes = NULL)

MemoryMappedFile$Read(nbytes = NULL)

BufferReader$Read(nbytes = NULL)

ReadableFile$Read(nbytes = NULL)

  

Arguments

nbytes

Number of bytes to read. Default NULL means to read all bytes.

Method ReadAt()

Similar to $seek(position)$Read(nbytes)

Usage

RandomAccessFile$ReadAt(position, nbytes = NULL)

MemoryMappedFile$ReadAt(position, nbytes = NULL)

BufferReader$ReadAt(position, nbytes = NULL)

ReadableFile$ReadAt(position, nbytes = NULL)

  

Arguments

position

Integer starting position in the stream

nbytes

Number of bytes to read. Default NULL means to read all bytes.

Method GetSize()

Usage

RandomAccessFile$GetSize()

MemoryMappedFile$GetSize()

BufferReader$GetSize()

ReadableFile$GetSize()

  

Value

An integer number of bytes that the stream has.

Method supports_zero_copy()

Usage

RandomAccessFile$supports_zero_copy()

MemoryMappedFile$supports_zero_copy()

BufferReader$supports_zero_copy()

ReadableFile$supports_zero_copy()

  

Value

Logical: whether the stream supports zero-copy access

Method seek()

Usage

RandomAccessFile$seek(position)

MemoryMappedFile$seek(position)

BufferReader$seek(position)

ReadableFile$seek(position)

  

Arguments

position

Integer starting position in the stream

Method tell()

Usage

RandomAccessFile$tell()

MemoryMappedFile$tell()

BufferReader$tell()

ReadableFile$tell()

  

Method Resize()

Usage

MemoryMappedFile$Resize(size)

Arguments

size

New size for the file